scripts/requirements.txt
pyyaml
dulwich